fix: 내 사진이면 좋아요 못 누르도록 고치기
